perf: 抽离公共复制方法；账号详情增加复制功能
