+++ src/utils/common.js
@@ -1,4 +1,5 @@
 import i18n from '@/i18n/i18n'
+import { message } from '@/utils/message'
 const _ = require('lodash')
 const moment = require('moment')
 
@@ -361,5 +362,20 @@ export function diffObject(object, base) {
     }
   })
 }
+
+export const copy = _.throttle(function(value) {
+  const inputDom = document.createElement('input')
+  inputDom.id = 'createInputDom'
+  inputDom.value = value
+  document.body.appendChild(inputDom)
+  inputDom.select()
+  document?.execCommand('copy')
+  message({
+    message: i18n.t('common.CopySuccess'),
+    type: 'success',
+    duration: 1000
+  })
+  document.body.removeChild(inputDom)
+}, 1400)
 
 export { BASE_URL }
